python/doclayout-yolo/requirements.txt
fastapi
uvicorn
pdf2image
opencv-python-headless
pillow
ultralytics
python-multipart
dill
doclayout_yolo
huggingface-hub


